+++ src/commands/polls/pollsButton.js
@@ -2,11 +2,12 @@ import { MessageActionRow, MessageSelectMenu } from "discord.js";
 import {
   fetchPollMessage,
   interactionEditReply,
-  refreshPollFields,
+  pollRefreshEmbed,
+  stopPoll,
 } from "./pollsUtils.js";
 import { createButton, isSentinelle } from "../utils.js";
 import { PERSONALITY } from "../../personality.js";
-import { getPoll, removePoll, resetPollVoters } from "../../helpers/index.js";
+import { getPoll, resetPollVoters } from "../../helpers/index.js";
 import { COMMONS } from "../../commons.js";
 
 export const sendSettingsButtons = async (interaction) => {
@@ -22,6 +23,12 @@ export const sendSettingsButtons = async (interaction) => {
   //check for Sentinelle or author
   const pollMessage = interaction.message;
   const dbPoll = getPoll(interaction.client.db, pollMessage.id);
+  if (!dbPoll) {
+    interactionEditReply(interaction, perso.errorNoPoll);
+    console.log(`Error poll not found : ${pollMessage.id}`);
+    return;
+  }
+
   if (interaction.user.id !== dbPoll.authorId) {
     //if not poll author, check is sentinelle
     const currentServer = COMMONS.fetchGuildId(interaction.guildId);
@@ -45,8 +52,8 @@ export const sendSettingsButtons = async (interaction) => {
 
   const firstButton = [updateButton, removeButton, resetButton, stopButton];
 
-  //If stoped poll, disable most buttons
-  if (pollEmbed.title.includes(perso.disable.title))
+  //If stopped poll, disable most buttons
+  if (pollEmbed.title.includes(perso.stop.title))
     firstButton.forEach((btn) => btn.setDisabled(true));
   const settingsButton = [refreshButton, ...firstButton];
 
@@ -57,41 +64,28 @@ export const sendSettingsButtons = async (interaction) => {
   interactionEditReply(interaction, { components: [actionRow] });
 };
 
-export const disablePoll = async (interaction) => {
+export const stopPollButtonAction = async (interaction) => {
   try {
     await interaction.deferUpdate();
   } catch (e) {
     return console.log(e);
   }
+  const db = interaction.client.db;
 
-  //get personality
+  //get data
   const perso = PERSONALITY.getCommands().polls.settings;
-
-  //fetch pollMessage
   const pollMessage = await fetchPollMessage(interaction);
-  const editedPollMessage = {};
+  const dbPoll = getPoll(db, pollMessage.id);
 
-  //edit title
-  const pollEmbed = pollMessage.embeds[0];
-  pollEmbed.title = pollEmbed.title + perso.disable.title;
-  editedPollMessage.embeds = [pollEmbed];
-
-  //edit poll buttons
-  const components = pollMessage.components;
-  const lastActionRow = components[components.length - 1];
-  const settingButton =
-    lastActionRow.components[lastActionRow.components.length - 1];
-  const newActionRow = new MessageActionRow().addComponents(settingButton);
-  editedPollMessage.components = [newActionRow];
+  await stopPoll(dbPoll, pollMessage, perso);
 
   //edit poll message
   const editedStopMessage = {
-    content: perso.disable.disabled,
+    content: perso.stop.stopped,
     components: [],
     ephemeral: true,
   };
   interactionEditReply(interaction, editedStopMessage);
-  pollMessage.edit(editedPollMessage);
 };
 
 export const removePollButtonAction = async (interaction) => {
@@ -207,9 +201,7 @@ export const refreshPollButtonAction = async (interaction) => {
   });
 
   const pollMessage = await fetchPollMessage(interaction);
-  const embed = pollMessage.embeds[0];
   const db = interaction.client.db;
-  const dbPoll = getPoll(db, pollMessage.id);
 
   //disable actions during refresh
   const disabledComponents = pollMessage.components.reduce((acc, cur) => {
@@ -218,15 +210,9 @@ export const refreshPollButtonAction = async (interaction) => {
   }, []);
   await pollMessage.edit({ components: disabledComponents });
 
-  //create new fields objects from pollMessage
-  const newFieldsInit = embed.fields.map((obj) => {
-    return { name: obj.name, value: "" };
-  }); //init with old names
-  const newFields = refreshPollFields(dbPoll, newFieldsInit, perso.create);
-
-  //update message
-  embed.setFields(newFields);
-  await pollMessage.edit({ embeds: [embed] });
+  //update poll embed
+  const dbPoll = getPoll(db, pollMessage.id);
+  await pollRefreshEmbed(pollMessage, dbPoll);
 
   //reply and enable votes
   interactionEditReply(interaction, {
@@ -234,16 +220,11 @@ export const refreshPollButtonAction = async (interaction) => {
     content: sPerso.refresh.done,
   });
 
-  //enable actions during refresh
-  if (!embed.title.includes(sPerso.disable.title)) {
-    //if not disabled, add again all buttons as enabled
-    const enabledComponents = pollMessage.components.reduce((acc, cur) => {
-      cur.components.forEach((button) => button.setDisabled(false)); //disable buttons of cur actionRow
-      return [...acc, cur];
-    }, []);
-    pollMessage.edit({ components: enabledComponents });
-  } else {
-    pollMessage.edit({ components: [] }); //else remove every button
-    removePoll(db, pollMessage.id); //remove data from db
-  }
+  //handle buttons
+  const enabledComponents = pollMessage.components.reduce((acc, cur) => {
+    cur.components.forEach((button) => button.setDisabled(false)); //disable buttons of cur actionRow
+    return [...acc, cur];
+  }, []);
+
+  pollMessage.edit({ components: enabledComponents });
 };